fix init without bson
